frameworks/pydantic-ai/lead-qual-agent-demo/pyproject.toml
[project]
name = "lead-qual-agent-demo"
version = "0.1.0"
description = "Lead qualification agent demo using Pydantic AI and Linkup.so"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Lead Qualification Agent Team", email = "team@example.com"},
]
dependencies = [
    "pydantic-ai>=0.0.9",
    "pydantic>=2.5.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "coverage>=7.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers -p no:logfire"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "C4", "DTZ", "T20", "SIM", "RET"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true